drift-analyzer/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gitpython==3.1.40
pyyaml==6.0.1
pydantic==2.5.0
python-multipart==0.0.6
